requirements.txt
pycryptodome==3.14.1
pyperclip==1.8.2
python-dotenv==0.19.2
